app/requirements.txt
numpy==1.21.5
Pillow==9.4.0
streamlit==1.17.0
tensorflow==2.11.0
